Almost there!
This fixes a lot of bugs with resizing and dragging, and adds a whole
bunch of default styling. As well as adding the resources to the left
side of the chart. It's looking awesome! Just need to add some callbacks
now and it should be ready for some real testing. The other thing that
needs to be tackled is dropping the event in the proper place when it
starts dragging from the middle of the event.