Fix scheduler
This fixes a lot of problems with the scheduler. Trying to reconcile
where an event was actually dropped when the dragging started in the
middle of an event proved to be super difficult. I updated the API so
that you can only drag from the event beginning. This removed A LOT of
complexity. I also added partial events so that events that spill over
into the next month are shown on the scheduler, although they are not
mutable unless they start in the current view. I also added a callback
to send out the new range when the date range is changed.