More performance tweaks and a bug fix.
This adds a bit more performance to the chart, though it still needs a
lot of work. There are way too many renders happening.

It also fixes a minor bug where events that spilled over to the left
side of the scheduler were covering up the resource name. This fixes
that using a zIndex and background color.